Consolidate track list appbar, add screen size variants.
